refactor: refact socket commit
